centralize button setup within button components and remove helper functions
