+++ client/one-time-payment/react/src/components/PayPalButton.tsx
@@ -1,19 +1,45 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext, useEffect, useRef } from "react";
 import { PayPalSDKContext } from "../context/sdkContext";
-import { setupPayPalButton } from "./utils";
+import { paymentSessionOptions, createOrder } from "./utils";
 
 const PayPalButton: React.FC = () => {
   const { isReady, sdkInstance, paymentMethodEligibility } =
     useContext(PayPalSDKContext);
+  const onClickRef = useRef(() => {});
 
   useEffect(() => {
     if (isReady && paymentMethodEligibility.isPayPalEligible) {
-      setupPayPalButton(sdkInstance);
+      const paypalSession = sdkInstance.createPayPalOneTimePaymentSession(
+        paymentSessionOptions
+      );
+      onClickRef.current = async () => {
+        try {
+          await paypalSession.start(
+            { presentationMode: "auto" },
+            createOrder()
+          );
+        } catch (e) {
+          console.error(e);
+        }
+      };
+
+      const paypalButton = document.getElementById("paypal-button");
+      paypalButton?.removeAttribute("hidden");
     }
   }, [isReady, paymentMethodEligibility.isPayPalEligible, sdkInstance]);
 
+  if (!isReady) {
+    return <p>LOADING.....</p>
+  }
 
-  return <paypal-button hidden type="pay" id="paypal-button"></paypal-button>;
+  return (
+    <paypal-button
+      onClick={() => onClickRef.current()}
+      hidden
+      type="pay"
+      id="paypal-button"
+    ></paypal-button>
+  );
 };
 
 export default PayPalButton;
